Created a new file lib/features/diary/widgets/appointment_dialog.dart containing the AppointmentDialog widget
