+++ lib/features/diary/tab.dart
@@ -30,6 +30,7 @@ import 'package:table_calendar/table_calendar.dart';
 import 'package:intl/intl.dart';
 
 import 'models/appointment.dart';
+import 'widgets/appointment_dialog.dart';
 
 class DiaryTab extends StatefulWidget {
   const DiaryTab({super.key});
@@ -106,7 +107,7 @@ class _DiaryTabState extends State<DiaryTab> {
   void _addAppointment() {
     showDialog(
       context: context,
-      builder: (context) => _AppointmentDialog(
+      builder: (context) => AppointmentDialog(
         onSave: (title, description, date) {
           setState(() {
             _appointments.add(
@@ -267,138 +268,6 @@ class _DiaryTabState extends State<DiaryTab> {
           ),
         ],
       ),
-    );
-  }
-}
-
-class _AppointmentDialog extends StatefulWidget {
-  final Function(String title, String description, DateTime date) onSave;
-
-  const _AppointmentDialog({required this.onSave});
-
-  @override
-  State<_AppointmentDialog> createState() => _AppointmentDialogState();
-}
-
-class _AppointmentDialogState extends State<_AppointmentDialog> {
-  final _titleController = TextEditingController();
-  final _descriptionController = TextEditingController();
-  DateTime _selectedDate = DateTime.now();
-  TimeOfDay _selectedTime = TimeOfDay.now();
-
-  @override
-  void dispose() {
-    _titleController.dispose();
-    _descriptionController.dispose();
-    super.dispose();
-  }
-
-  Future<void> _selectDate() async {
-    final DateTime? picked = await showDatePicker(
-      context: context,
-      initialDate: _selectedDate,
-      firstDate: DateTime.now(),
-      lastDate: DateTime.now().add(const Duration(days: 365)),
-    );
-    if (picked != null) {
-      setState(() {
-        _selectedDate = DateTime(
-          picked.year,
-          picked.month,
-          picked.day,
-          _selectedTime.hour,
-          _selectedTime.minute,
-        );
-      });
-    }
-  }
-
-  Future<void> _selectTime() async {
-    final TimeOfDay? picked = await showTimePicker(
-      context: context,
-      initialTime: _selectedTime,
-    );
-    if (picked != null) {
-      setState(() {
-        _selectedTime = picked;
-        _selectedDate = DateTime(
-          _selectedDate.year,
-          _selectedDate.month,
-          _selectedDate.day,
-          picked.hour,
-          picked.minute,
-        );
-      });
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      title: const Text('Add Appointment'),
-      content: SingleChildScrollView(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            TextField(
-              controller: _titleController,
-              decoration: const InputDecoration(
-                labelText: 'Title',
-                border: OutlineInputBorder(),
-              ),
-            ),
-            const SizedBox(height: 16),
-            TextField(
-              controller: _descriptionController,
-              decoration: const InputDecoration(
-                labelText: 'Description',
-                border: OutlineInputBorder(),
-              ),
-              maxLines: 3,
-            ),
-            const SizedBox(height: 16),
-            Row(
-              children: [
-                Expanded(
-                  child: OutlinedButton.icon(
-                    onPressed: _selectDate,
-                    icon: const Icon(Icons.calendar_today),
-                    label: Text(
-                      DateFormat('MMM dd, yyyy').format(_selectedDate),
-                    ),
-                  ),
-                ),
-                const SizedBox(width: 8),
-                Expanded(
-                  child: OutlinedButton.icon(
-                    onPressed: _selectTime,
-                    icon: const Icon(Icons.access_time),
-                    label: Text(_selectedTime.format(context)),
-                  ),
-                ),
-              ],
-            ),
-          ],
-        ),
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.pop(context),
-          child: const Text('Cancel'),
-        ),
-        TextButton(
-          onPressed: () {
-            if (_titleController.text.isNotEmpty) {
-              widget.onSave(
-                _titleController.text,
-                _descriptionController.text,
-                _selectedDate,
-              );
-            }
-          },
-          child: const Text('Save'),
-        ),
-      ],
     );
   }
 }
